feat: port ramp_30 modifications from REISE.jl
